Added code to potentially fix the overwriting existing partitions error, still needs testing

+++ finish_install.sh
@@ -96,18 +96,20 @@ echo '/swapfile none swap 0 0' >> /etc/fstab
 #----------------  Grub Configuration ----------------
 clear
 
-uefi_enabled=`cat uefi_state.temp`
+uefi_enabled=$(cat uefi_state.temp)
 
-disk_label=`cat disk_label.temp`
+disk_label=$(cat disk_label.temp)
+disk_number=$(cat disk_number.temp)
+system_partition=$(cat next_open_partition.temp)
+boot_partition=$(cat boot_partition.temp)
+existing_boot_partition=$(cat existing_boot_partition.temp)
 
-disk_number=`cat disk_number.temp`
-
-encrypt_system=`cat encrypted_system.temp`
+encrypt_system=$(cat encrypted_system.temp)
 
 if [ $encrypt_system == 'y' ] || [ $encrypt_system == 'Y' ] || [ $encrypt_system == 'yes' ]
 then
   # Encryption configuration
-  echo "GRUB_CMDLINE_LINUX='cryptdevice=/dev/${disk_number}2:cryptdisk'" >> /etc/default/grub
+  echo "GRUB_CMDLINE_LINUX='cryptdevice=${system_partition}:cryptdisk'" >> /etc/default/grub
   echo -e 'MODULES=()\nBINARIES=()\nFiles=()\nHOOKS=(base udev autodetect modconf block encrypt filesystems keyboard fsck)' > /etc/mkinitcpio.conf
 fi
 
@@ -116,13 +118,17 @@ echo -e '\nGRUB_DISABLE_OS_PROBER=false\nGRUB_SAVEDEFAULT=true\nGRUB_DEFAULT=sav
 pacman -S --noconfirm linux linux-lts
 mkinitcpio --allpresets
 
-# Actual Grub Install
-if [ $uefi_enabled == True ]
+# Only install grub if a boot partition doesn't already exist
+if [[ $existing_boot_partition != True ]];
 then
-  pacman -S --noconfirm efibootmgr dosfstools mtools
-  grub-install --efi-directory=/boot
-else
-  grub-install /dev/${disk_label}
+    # Actual Grub Install
+    if [ $uefi_enabled == True ]
+    then
+      pacman -S --noconfirm efibootmgr dosfstools mtools
+      grub-install --efi-directory=/boot
+    else
+      grub-install /dev/${disk_label}
+    fi
 fi
 grub-mkconfig -o /boot/grub/grub.cfg
 
@@ -136,6 +142,10 @@ rm uefi_state.temp
 rm disk_label.temp
 rm disk_number.temp
 rm finish_install.sh
+rm next_open_partition.temp
+rm boot_partition.temp
+rm existing_boot_partition.temp
+
 
 clear
 exit
